Use newtypes everywhere
The low-level bindings are leaking far more than is necessary, in ways
that can lead to subtle bugs because of insufficient abstraction
boundaries. Newtypes are cheap and should be used nearly always in
Haskell. Type synonyms should not be used for abstraction, only
abbreviation. Chris Done articulates this point well:

https://www.reddit.com/r/haskell/comments/64ui11/reasons_i_dislike_type_synonyms/