Design bindings to be imported qualified
All names in the C API exported by Binaryen start with the `Binaryen`
prefix. This prefix is both unnecessary and problematic, because it's
capitalized, which in Haskell doesn't work in the case of functions.
This was forcing us previously to prefix each function with `c_`,
which presumably higher level bindings would remove. Except that there
are many functions, so the cost of having both low-level and separate
high-level bindings is high.

We now drop the `Binaryen` prefix, both for types and for functions.