Split up monolithic binding module
Upstream ships the API as a single gigantic header file. Since this is
a C API, there is no namespacing. That doesn't mean we need to do the
same. We instead design the bindings with the idea that modules will
be imported qualified by the user. This makes it easier to group
related functions together, it's easier to manage multiple smaller
modules rather than one large module, and we can shorten function names.